+++ device_info_plugin_ios/example/device_info_plugin_ios_example.dart
@@ -1,6 +1,0 @@
-import 'package:device_info_plugin_ios/device_info_plugin_ios.dart';
-
-void main() {
-  var awesome = Awesome();
-  print('awesome: ${awesome.isAwesome}');
-}