+++ device_info_plugin_ios/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.0.4+1
+
+ - **REFACTOR**(ios): rename plugin classes and add example app. ([d2e3f749](https://github.com/Thixq/device_info_plugin.git/commit/d2e3f7490b6fd5e7e412f17e00f799d4ab48773a))
+
+## 0.0.4
+
+ - **FEAT**(ios): implement native iOS plugin using Pigeon, add tests, and update pubspec. ([7e44e087](https://github.com/Thixq/device_info_plugin.git/commit/7e44e08779f212ac0589d37d9c7d74eb56580bfa))
+ - **FEAT**(ios): setup native ios plugin structure and generate pigeon models. ([299ae008](https://github.com/Thixq/device_info_plugin.git/commit/299ae008c57b80212cb65a4d8787b78eab22179f))
+
 ## 0.0.3
 
  - **FEAT**(ios): configure pigeon and migrate to very_good_analysis. ([4d2c4189](https://github.com/Thixq/device_info_plugin.git/commit/4d2c4189968a593a42f67faf8dc4c62510aee95d))